컴파일 빌드 수정 (패키지 및 라이브러리 재설정)

+++ src/main/java/com/libqa/repository/WikiRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by yong on 15. 2. 8..
  */
-public interface WikiRepository extends JpaRepository<Wiki, Long> {
+public interface WikiRepository extends JpaRepository<Wiki, Integer> {
 }
